# src/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SSI API",
    "description": "Self-Sovereign Identity API for management of PKK & Trusted Lists"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/pki/pkd": {
      "get": {
        "tags": [
          "Public Key Directory"
        ],
        "summary": "List PKDs",
        "description": "List all deployed Public Key Directories",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": [
                "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621A",
                "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621B"
              ]
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/pkd/{address}/entities": {
      "get": {
        "tags": [
          "Public Key Directory"
        ],
        "summary": "List the entities of a PKD",
        "description": "List all the entities registered in the specified Public Key Directory",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "The PKD address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621B"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": [
                {
                  "address": "Entity address",
                  "did": "Entity DID",
                  "expires": "Entity expiration date",
                  "status": "Entity status (active|revoked)"
                }
              ]
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/pkd/deploy": {
      "post": {
        "tags": [
          "Public Key Directory"
        ],
        "summary": "Deploy PKD",
        "description": "Deploy a new Public Key Directory contract",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": {
                "address": "The PKD address",
                "hash": "The blockchain transaction hash"
              }
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/pkd/{address}/register": {
      "post": {
        "tags": [
          "Public Key Directory"
        ],
        "summary": "Register an entity in the PKD",
        "description": "Register a new entity in the Public Key Directory",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "The PKD address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621B"
          },
          {
            "name": "body",
            "description": "The entity information",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "description": "The entity address (useful when is different from DID, for example TL address)",
                  "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621A"
                },
                "did": {
                  "type": "string",
                  "description": "The DID of the entity",
                  "default": "did:lac:main:0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621A"
                },
                "expires": {
                  "type": "string",
                  "description": "The expiration date in UNIX time - seconds",
                  "default": "1757247663"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": {
                "hash": "The blockchain transaction hash"
              }
            }
          },
          "404": {
            "description": "PKD not found or invalid address"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/pkd/{address}/revoke/{entity}": {
      "delete": {
        "tags": [
          "Public Key Directory"
        ],
        "summary": "Revoke an entity in the PKD",
        "description": "Revoke an entity registered in the Public Key Directory",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "The PKD address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621B"
          },
          {
            "name": "entity",
            "description": "The Entity address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621C"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": {
                "hash": "The blockchain transaction hash"
              }
            }
          },
          "404": {
            "description": "PKD not found or invalid address"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/tl": {
      "get": {
        "tags": [
          "Trusted List"
        ],
        "summary": "List TLs",
        "description": "List all deployed Trusted Lists",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": [
                {
                  "address": "The entity address",
                  "parent": "The address of the parent TL or PKD",
                  "name": "The name of the entity"
                }
              ]
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/tl/{address}/entities": {
      "get": {
        "tags": [
          "Trusted List"
        ],
        "summary": "List the entities of a TL",
        "description": "List all the entities registered in the specified Trusted List",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "The TL address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621B"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": [
                {
                  "address": "Entity address",
                  "did": "Entity DID",
                  "expires": "Entity expiration date",
                  "status": "Entity status (active|revoked)"
                }
              ]
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/tl/deploy": {
      "post": {
        "tags": [
          "Trusted List"
        ],
        "summary": "Deploy TL",
        "description": "Deploy a new Trusted List contract",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "description": "The entity information",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "parent": {
                  "type": "string",
                  "description": "The parent (TL or PKD) address",
                  "default": "0xF49ECE7f8aaCfba94EB5f7b296Ad7175d9BB501d"
                },
                "name": {
                  "type": "string",
                  "description": "The name of the entity",
                  "default": "Trusted List 1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": {
                "address": "The TL address",
                "hash": "The blockchain transaction hash"
              }
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/tl/{address}/register": {
      "post": {
        "tags": [
          "Trusted List"
        ],
        "summary": "Register an entity in the TL",
        "description": "Register a new entity in the Public Key Directory",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "The TL address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D6313"
          },
          {
            "name": "body",
            "description": "The entity information",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "description": "The entity address (useful when is different from DID, for example TL address)",
                  "default": "0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D671A"
                },
                "did": {
                  "type": "string",
                  "description": "The DID of the entity",
                  "default": "did:lac:main:0x5967b9b16024Ffc06Ef25Af6D77de00c5C5D621A"
                },
                "name": {
                  "type": "string",
                  "description": "The name of the entity",
                  "default": "Entity 1 Inc."
                },
                "expires": {
                  "type": "number",
                  "description": "The expiration date in UNIX time - seconds",
                  "default": 1757247663
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": {
                "hash": "The blockchain transaction hash"
              }
            }
          },
          "404": {
            "description": "TL not found or invalid address"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
    "/pki/tl/{address}/revoke/{entity}": {
      "delete": {
        "tags": [
          "Trusted List"
        ],
        "summary": "Revoke an entity in the TL",
        "description": "Revoke an entity registered in the Trusted List",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "description": "The TL address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x6967b9b16024Ffc06Ef25Af6D77de00c5C5D621B"
          },
          {
            "name": "entity",
            "description": "The Entity address",
            "in": "path",
            "required": true,
            "type": "string",
            "default": "0x6967b9b16024Ffc06Ef25Af6D77de00c5C5D621C"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "examples": {
              "application/json": {
                "hash": "The blockchain transaction hash"
              }
            }
          },
          "404": {
            "description": "TL not found or invalid address"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    }
  }
}